fix: 전역 폰트 모달에서 적용 안되는 오류 해결
